changes from last lesson

+++ macros/delete_from.sql
@@ -1,3 +1,8 @@
 {% macro delete_from(table_name) %}
-    delete from {{table_name}};
+
+    {% if is_incremental() %}
+      delete from {{table_name}};
+    {% endif %}
+
+
 {% endmacro %}
